Moved old code to its own packages, robot.argos, robot.bigwheelreplay, robot.kraken
