+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robots/kraken/KrakenLEDSystem.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.robots.kraken;
 
 /**
  * Created by Karim on 3/9/2018.
