+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/robots/kraken/PoseKraken.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.robots.kraken;
 
 import android.util.Log;
 
@@ -15,6 +15,8 @@ import org.firstinspires.ftc.robotcore.external.navigation.AxesOrder;
 import org.firstinspires.ftc.robotcore.external.navigation.AxesReference;
 import org.firstinspires.ftc.robotcore.external.navigation.Orientation;
 import org.firstinspires.ftc.robotcore.external.navigation.VuforiaTrackableDefaultListener;
+import org.firstinspires.ftc.teamcode.PIDController;
+import org.firstinspires.ftc.teamcode.RC;
 
 
 /**
@@ -29,7 +31,7 @@ import org.firstinspires.ftc.robotcore.external.navigation.VuforiaTrackableDefau
  * @since 2016-12-10
  */
 
-public class PoseKrakatoa
+public class PoseKraken
 {
 
     HardwareMap hwMap;
@@ -62,7 +64,7 @@ public class PoseKrakatoa
     Servo servoLiftRight             = null;
     Servo servoPhone                 = null;
     Servo servoJewelFlipper = null;
-    Servo cannon                     = null;
+    Servo servoLED                   = null;
     Servo relicShoulder              = null;
     Servo relicElbow                 = null;
     Servo relicGrip                  = null;
@@ -118,14 +120,11 @@ public class PoseKrakatoa
     private double poseSavedHeading = 0.0;
     int balanceState = 1;
 
-    public boolean isCannonOn = false;
-    public double cannonOn = 1;
-    public double cannonOff = .5;
-
     //scoring objects and related variables
 //    public GlyphSystem glyphSystem;
     public JewelArm jewel;
     public GlyphSystem2 glyphSystem;
+    public KrakenLEDSystem ledSystem = null;
     public RelicArm relicArm = null;
 
 
@@ -189,7 +188,7 @@ public class PoseKrakatoa
      * @param heading The heading of the robot
      * @param speed The speed of the robot
      */
-    public PoseKrakatoa(double x, double y, double heading, double speed)
+    public PoseKraken(double x, double y, double heading, double speed)
     {
 
         poseX     = x;
@@ -208,7 +207,7 @@ public class PoseKrakatoa
      * @param y     The position relative to the y axis of the field
      * @param angle The vuAngle of the robot
      */
-    public PoseKrakatoa(double x, double y, double angle)
+    public PoseKraken(double x, double y, double angle)
     {
 
         poseX     = x;
@@ -222,7 +221,7 @@ public class PoseKrakatoa
      * Creates a base Pose instance at the origin, (_0,_0), with _0 speed and _0 vuAngle.
      * Useful for determining the Pose of the robot relative to the origin.
      */
-    public PoseKrakatoa()
+    public PoseKraken()
     {
 
         poseX     = 0;
@@ -279,7 +278,7 @@ public class PoseKrakatoa
         this.motorGripLeft      = this.hwMap.dcMotor.get("motorGripLeft");
         this.motorGripRight     = this.hwMap.dcMotor.get("motorGripRight");
         this.servoJewelFlipper  = this.hwMap.servo.get("servoJewelFlipper");
-        this.cannon             = this.hwMap.servo.get("cannon");
+        this.servoLED           = this.hwMap.servo.get("servoLED");
         this.relicShoulder      = this.hwMap.servo.get("relicShoulder");
         this.relicElbow         = this.hwMap.servo.get("relicElbow");
         this.relicGrip          = this.hwMap.servo.get("relicGrip");
@@ -288,6 +287,8 @@ public class PoseKrakatoa
         //this.glyphUpper       = this.hwMap.get(DistanceSensor.class, "glyphDetectUp");
 
         //motor configurations
+
+        this.ledSystem=new KrakenLEDSystem(this.servoLED);
 
         this.motorFrontRight.setDirection(DcMotorSimple.Direction.REVERSE);
         this.motorBackRight.setDirection(DcMotorSimple.Direction.REVERSE);
@@ -330,18 +331,6 @@ public class PoseKrakatoa
         headLampOn();
     }
 
-    public void cannonOn(){
-        isCannonOn = true;
-    }
-
-    public void cannonOff(){
-        isCannonOn = false;
-    }
-
-    public void toggleCannon(){
-        isCannonOn = !isCannonOn;
-    }
-
 
     /**
      * update the current location of the robot. This implementation gets heading and orientation
@@ -379,9 +368,6 @@ public class PoseKrakatoa
         }
 
         glyphSystem.update();
-
-        if(isCannonOn) cannon.setPosition(cannonOn);
-        else  cannon.setPosition(cannonOff);
 
         poseHeading = wrapAngle(imuAngles.firstAngle, offsetHeading);
         posePitch = wrapAngle(imuAngles.thirdAngle, offsetPitch);
