+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/vision/colorblob/ColorBlobDetector.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.vision.colorblob;
 
 import org.firstinspires.ftc.teamcode.util.BlobStats;
 import org.opencv.core.Core;
